+++ src/Examples/ExampleWpfApplication/BooleanToVisibilityConverter.cs
@@ -4,13 +4,13 @@ using System.Windows;
 using System.Windows.Data;
 using FeatureToggle.Core;
 
-namespace FeatureToggle.Toggles
-{    
-    public class FeatureToggleToVisibilityConverter : IValueConverter
-    {   
+namespace ExampleWpfApplication
+{
+    class BooleanToVisibilityConverter : IValueConverter
+    {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            var toggle = (IFeatureToggle)value;
+            var toggle = (IFeatureToggle) value;
 
             return toggle.FeatureEnabled ? Visibility.Visible : Visibility.Collapsed;
         }