+++ Release-Notes.md
@@ -7,6 +7,7 @@ Currently in development.
 ##Breaking Changes
 
 * Remove the WPF, etc Visibility converters. [#67](https://github.com/jason-roberts/FeatureToggle/issues/67)
+* Remove Windows phone 8.0 support [#80](https://github.com/jason-roberts/FeatureToggle/issues/80)
 
 ## Enhancements / New Features